Added common module with useful strategies from hypothesis. Added conftest file so that all test folder is imported by pytest
